% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzmem-package.R
\docType{package}
\name{gzmem}
\alias{gzmem}
\alias{gzmem-package}
\title{Better gz in memory compression/decompression}
\description{
A partial resurrection of `Rcompression` for
\href{http://stackoverflow.com/q/39707388/1457051}{this}.
}
\author{
Bob Rudis (@hrbrmstr)
}

