% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mem_compress}
\alias{mem_compress}
\title{Compress a raw vector using gz compression}
\usage{
mem_compress(r_content)
}
\arguments{
\item{r_content}{raw vector of data to compress}
}
\value{
Raw vector of compressed data
}
\description{
Compress a raw vector using gz compression
}
\examples{
rawToChar(mem_inflate(mem_compress(
  charToRaw("The quick brown fox jumps over the lazy dog.")), 1000))
}

