% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mem_inflate}
\alias{mem_inflate}
\title{Inflate a raw vector that was compressed}
\usage{
mem_inflate(r_source, format, r_guess_size)
}
\arguments{
\item{r_source}{raw vector of compressed data.}

\item{format}{A character scalar indicating the type of compression that was applied.
This must be one of "gzip", "zlib" or "raw".}

\item{r_guess_size}{your best guess as to the size of uncompressed data. Not ideal,
and this won't be necessary in future releases. Reember, this is a direct
port from the defunct \code{Rcompression} package.\cr
\cr
Aim high as you'll only get back the actual number of bytes
in the uncompressed data.}
}
\value{
Raw vector of expanded data.
}
\description{
Inflate a raw vector that was compressed
}
\examples{
raw <- charToRaw("The quick brown fox jumps over the lazy dog.")
compressed <- mem_compress(raw, format = "gzip")
decompressed <- mem_inflate(compressed, format = "gzip", 1000)
}
