% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mem_inflate}
\alias{mem_inflate}
\title{Inflate a raw vector that was gz compressed}
\usage{
mem_inflate(r_source, r_guess_size)
}
\arguments{
\item{r_source}{raw vector of gz compressed data}

\item{r_guess_size}{your best guess as to the size of uncompressed data. Not ideal,
and this won't be necessary in future releases. Reember, this is a direct
port from the defunct \code{Rcompression} package.\cr
\cr
Aim high as you'll only get back the actual number of bytes
in the uncompressed data.}

\item{Raw}{vector of expanded data}
}
\description{
Inflate a raw vector that was gz compressed
}
\examples{
rawToChar(mem_inflate(mem_compress(
  charToRaw("The quick brown fox jumps over the lazy dog.")), 1000))
}

